improve api

+++ controllers/bookings.controller.js
@@ -13,6 +13,10 @@ controller.getBookings = async (req, res) => {
 
 controller.deleteBooking = async (req, res) => {
     try {
+        var checkForHexRegExp = new RegExp("^[0-9a-fA-F]{24}$")
+        if(!checkForHexRegExp.test(req.params.element)) {
+            return res.status(404).send("Invalid URL");
+        }
         await Booking.findByIdAndDelete(req.params.element)
         return res.send("Sucessfully deleted")
     } catch(e) {
@@ -23,6 +27,10 @@ controller.deleteBooking = async (req, res) => {
 
 controller.editBooking = async (req, res) => {
     try {
+        var checkForHexRegExp = new RegExp("^[0-9a-fA-F]{24}$")
+        if(!checkForHexRegExp.test(req.params.element)) {
+            return res.status(404).send("Invalid URL");
+        }
       let response = await Booking.findByIdAndUpdate( req.params.element, req.body, (e) => {
           if (e) throw e;
       }).clone();
@@ -35,6 +43,10 @@ controller.editBooking = async (req, res) => {
 
 controller.searchBooking = async (req, res) => {
     try {
+        var checkForHexRegExp = new RegExp("^[0-9a-fA-F]{24}$")
+        if(!checkForHexRegExp.test(req.params.element)) {
+            return res.status(404).send("Invalid URL");
+        }
         let response = await Booking.findById(req.params.element, (e) =>{
             if (e) throw e;
         }).clone();
